+++ CHANGELOG.md
@@ -2,3 +2,6 @@ v1.0.2 -> v1.0.3
 
 - added highlight color option
 - added more backgrounds
+- added more animation types
+- added ui background and ui border color customizability
+- added main text color